requirements.txt
google-cloud-storage
absl-py
json_lines
ipython
https://storage.googleapis.com/jax-releases/cuda100/jaxlib-0.1.51-cp37-none-manylinux2010_x86_64.whl
jax
tfds-nightly
tensorflow
tensorflow-text
tqdm
msgpack>=1.0.0
--editable git+https://github.com/google-research/reverse-engineering-neural-networks.git@master#egg=renn
--editable git+https://github.com/google/uv-metrics.git@master#egg=uv-metrics
pg8000>=1.16.1
